Push to production

+++ docs-web/src/main/java/com/sismics/docs/rest/resource/AclResource.java
@@ -1,39 +1,26 @@
 package com.sismics.docs.rest.resource;
 
-import java.text.MessageFormat;
-import java.util.List;
-
-import javax.json.Json;
-import javax.json.JsonArrayBuilder;
-import javax.json.JsonObjectBuilder;
-import javax.ws.rs.DELETE;
-import javax.ws.rs.FormParam;
-import javax.ws.rs.GET;
-import javax.ws.rs.PUT;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.QueryParam;
-import javax.ws.rs.core.Response;
-
 import com.google.common.collect.Lists;
 import com.sismics.docs.core.constant.AclTargetType;
 import com.sismics.docs.core.constant.PermType;
-import com.sismics.docs.core.dao.jpa.AclDao;
-import com.sismics.docs.core.dao.jpa.DocumentDao;
-import com.sismics.docs.core.dao.jpa.GroupDao;
-import com.sismics.docs.core.dao.jpa.UserDao;
+import com.sismics.docs.core.dao.jpa.*;
 import com.sismics.docs.core.dao.jpa.criteria.GroupCriteria;
 import com.sismics.docs.core.dao.jpa.criteria.UserCriteria;
 import com.sismics.docs.core.dao.jpa.dto.GroupDto;
 import com.sismics.docs.core.dao.jpa.dto.UserDto;
-import com.sismics.docs.core.model.jpa.Acl;
-import com.sismics.docs.core.model.jpa.Document;
-import com.sismics.docs.core.model.jpa.Group;
-import com.sismics.docs.core.model.jpa.User;
+import com.sismics.docs.core.model.jpa.*;
 import com.sismics.docs.core.util.jpa.SortCriteria;
 import com.sismics.rest.exception.ClientException;
 import com.sismics.rest.exception.ForbiddenClientException;
 import com.sismics.rest.util.ValidationUtil;
+
+import javax.json.Json;
+import javax.json.JsonArrayBuilder;
+import javax.json.JsonObjectBuilder;
+import javax.ws.rs.*;
+import javax.ws.rs.core.Response;
+import java.text.MessageFormat;
+import java.util.List;
 
 /**
  * ACL REST resources.
@@ -48,7 +35,7 @@ public class AclResource extends BaseResource {
      * @param sourceId Source ID
      * @param permStr Permission
      * @param targetName Target name
-     * @param type ACL type
+     * @param typeStr ACL type
      * @return Response
      */
     @PUT
@@ -156,7 +143,14 @@ public class AclResource extends BaseResource {
         if (document != null && document.getUserId().equals(targetId)) {
             throw new ClientException("AclError", "Cannot delete base ACL on a document");
         }
-        
+
+        // Cannot delete R/W on a source tag if the target is the creator
+        TagDao tagDao = new TagDao();
+        Tag tag = tagDao.getById(sourceId);
+        if (tag != null && tag.getUserId().equals(targetId)) {
+            throw new ClientException("AclError", "Cannot delete base ACL on a tag");
+        }
+
         // Delete the ACL
         aclDao.delete(sourceId, perm, targetId, principal.getId());
